@@ -3,30 +3,28 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@craco/craco": "^6.2.0",
-    "@heroicons/react": "^1.0.3",
-    "@reduxjs/toolkit": "^1.6.1",
-    "@tailwindcss/forms": "^0.3.3",
-    "@tailwindcss/line-clamp": "^0.2.1",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.21.1",
-    "date-fns": "^2.24.0",
-    "moment": "^2.29.1",
-    "react": "^17.0.2",
-    "react-country-region-selector": "^3.3.0",
-    "react-dom": "^17.0.2",
+    "@craco/craco": "^6.4.0",
+    "@heroicons/react": "^2.0.8",
+    "@reduxjs/toolkit": "^1.8.5",
+    "@testing-library/user-event": "^14.4.3",
+    "axios": "^0.27.2",
+    "date-fns": "^2.29.2",
+    "moment": "^2.29.4",
+    "react": "^18.2.0",
+    "react-country-region-selector": "^3.6.1",
+    "react-dom": "^18.2.0",
+    "react-markdown": "^8.0.3",
     "react-phone-number-input": "^3.1.25",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-select": "^4.3.1",
-    "react-slider": "^1.3.1",
-    "react-table": "^7.7.0",
-    "redux": "^4.1.1",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "react-select": "^5.4.0",
+    "react-slider": "^2.0.3",
+    "react-syntax-highlighter": "^15.5.0",
+    "redux": "^4.2.0",
     "timeago.js": "^4.0.2",
-    "web-vitals": "^1.1.2"
+    "typescript": "^4.8.2",
+    "web-vitals": "^3.0.0"
   },
   "scripts": {
     "start": "craco start",
@@ -53,8 +51,16 @@
     ]
   },
   "devDependencies": {
-    "autoprefixer": "^9.8.6",
-    "postcss": "^7.0.36",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.7"
+    "@tailwindcss/forms": "^0.5.2",
+    "@tailwindcss/line-clamp": "^0.4.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.3.0",
+    "@types/jest": "^28.1.8",
+    "@types/node": "^18.7.13",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
+    "autoprefixer": "^10.4.8",
+    "postcss": "^8.4.16",
+    "tailwindcss": "3.1.8"
   }
 }
