@@ -1,13 +1,19 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
+import ReactDOM from 'react-dom/client';
 import App from './App';
+import './index.css';
 import { Provider } from 'react-redux';
-import configureStore from './store/configureStore'
+import configureStore from './store/configureStore';
+import { StrictMode } from 'react';
+import { BrowserRouter } from 'react-router-dom';
 const store = configureStore();
-ReactDOM.render(
-  <Provider store={store}>
-    <App />
-  </Provider>,
-  document.getElementById('root')
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <StrictMode>
+    <BrowserRouter>
+      <Provider store={store}>
+        <App />
+      </Provider>
+    </BrowserRouter>
+  </StrictMode>
 );
