@@ -1,10 +1,7 @@
 module.exports = {
-    style: {
-      postcss: {
-        plugins: [
-          require('tailwindcss'),
-          require('autoprefixer'),
-        ],
-      },
+  style: {
+    postcssOptions: {
+      plugins: [require('tailwindcss'), require('autoprefixer')],
     },
-  }
+  },
+};
