@@ -1,9 +1,13 @@
 import axios from '../../axios';
+import { logout } from '../user/api.user';
 
 import {
   ENDORSE_USER_REQUEST,
   ENDORSE_USER_SUCCESS,
   ENDORSE_USER_FAIL,
+  ENDORSE_DELETE_REQUEST,
+  ENDORSE_DELETE_SUCCESS,
+  ENDORSE_DELETE_FAIL,
   GET_GIVEN_ENDORSEMENT_REQUEST,
   GET_GIVEN_ENDORSEMENT_SUCCESS,
   GET_GIVEN_ENDORSEMENT_FAIL,
@@ -109,68 +113,40 @@ export const endorseUser = (endorsement) => async (dispatch, getState) => {
   }
 };
 
-// export const applicationDetail = (id) => async (dispatch, getState) => {
-//   try {
-//     dispatch({
-//       type: APPLICATION_DETAIL_REQUEST,
-//     });
-//     const {
-//       userLogin: { currentUser },
-//     } = getState();
+export const deleteEndorsement = (id) => async (dispatch, getState) => {
+  try {
+    dispatch({
+      type: ENDORSE_DELETE_REQUEST,
+    });
 
-//     const config = {
-//       headers: {
-//         Authorization: `Bearer ${currentUser.token}`,
-//       },
-//     };
-//     const { data } = await axios.get(`/applications/${id}`, config);
-//     dispatch({
-//       type: APPLICATION_DETAIL_SUCCESS,
-//       payload: data,
-//     });
-//   } catch (error) {
-//     dispatch({
-//       type: APPLICATION_DETAIL_FAIL,
-//       payload:
-//         error.response && error.response.data.message
-//           ? error.response.data.message
-//           : error.message,
-//     });
-//   }
-// };
+    const {
+      userLogin: { currentUser },
+    } = getState();
 
-// export const updateApplication =
-//   (id, auditionPostId, updateData) => async (dispatch, getState) => {
-//     try {
-//       dispatch({
-//         type: APPLICATION_UPDATE_REQUEST,
-//       });
-//       const {
-//         userLogin: { currentUser },
-//       } = getState();
+    const config = {
+      headers: {
+        Authorization: `Bearer ${currentUser.token}`,
+      },
+    };
 
-//       const config = {
-//         headers: {
-//           Authorization: `Bearer ${currentUser.token}`,
-//         },
-//       };
-//       const { data } = await axios.put(
-//         `/applications/${id}`,
-//         updateData,
-//         config
-//       );
-//       dispatch({
-//         type: APPLICATION_UPDATE_SUCCESS,
-//         payload: data,
-//       });
-//       dispatch(listApplications(auditionPostId));
-//     } catch (error) {
-//       dispatch({
-//         type: APPLICATION_UPDATE_FAIL,
-//         payload:
-//           error.response && error.response.data.message
-//             ? error.response.data.message
-//             : error.message,
-//       });
-//     }
-//   };
+    const { data } = await axios.delete(`/endorsements/${id}`, config);
+
+    dispatch({
+      type: ENDORSE_DELETE_SUCCESS,
+      payload: data,
+    });
+    dispatch(getGivenEndorsements());
+  } catch (error) {
+    const message =
+      error.response && error.response.data.message
+        ? error.response.data.message
+        : error.message;
+    if (message === 'Not authorized, no token') {
+      dispatch(logout ());
+    }
+    dispatch({
+      type: ENDORSE_DELETE_FAIL,
+      payload: message,
+    });
+  }
+};
