@@ -1,6 +1,7 @@
 module.exports = {
   purge: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
-  darkMode: false, // or 'media' or 'class'
+  darkMode: false,
+  // or 'media' or 'class'
   theme: {
     fontSize: {
       xs: ['.6rem', '.75rem'],
